Float support changed.

+++ test/proc.c
@@ -21,20 +21,20 @@ int proc02 ( int *io_par1, int *o_par2, int i_par3 )
 
 double proc03 ( double *o_par1, double i_par2, double *io_par3 )
 {
-	printf( "proc03\ni_par2: %lf\nio_par3: %lf\n", i_par2, *io_par3 );
+	printf( "proc03\ni_par2: %f\nio_par3: %f\n", i_par2, *io_par3 );
 	*o_par1 = 100.9998;
 	*io_par3 = 13478.54754;
-	printf( "o_par1: %lf\no_par3: %lf\n\n", *o_par1, *io_par3 );
+	printf( "o_par1: %f\no_par3: %f\n\n", *o_par1, *io_par3 );
 	return 3473.686;
 }
 
 float proc04 ( float i_par1, float *io_par2, float *o_par3 )
 {
-	printf( "proc04\ni_par1: %f; io_par2: %f\n", i_par1, *io_par2 );
-	*io_par2 = 33.99;
-	*o_par3 = -1.88;
-	printf( "proc04\nio_par2: %f; o_par3: %f\n", *io_par2, *o_par3 );
-	return 33.45;
+	printf( "proc04\ni_par1: %f; io_par2: %f\n", (double)i_par1, (double)*io_par2 );
+	*io_par2 = 33.99F;
+	*o_par3 = -1.88F;
+	printf( "proc04\nio_par2: %f; o_par3: %f\n", (double)*io_par2, (double)*o_par3 );
+	return 33.45F;
 }
 
 void nothing( void )
